restructuring super-admin/views
