+++ src/admin_portal/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from . import views
+from .views import views
 
 urlpatterns = [
   path('', views.test, name='test'),
